Backend/Application_Service/requirements.txt
django
djangorestframework
mongoengine
django-rest-framework-mongoengine

#  python -m pip install -r .\requirements.txt